Refactor code on Product Module
Refactor and optimization code on service, controller, models and dto's of products. Limit users on 'ProductsController', sending 403 code, if the user get into a incorrect endpoint.
