+++ API_Shopping/API_Shopping/Interfaces/IProductService.cs
@@ -1,11 +1,12 @@
-﻿using API_Shopping.Models;
+﻿using API_Shopping.DTOs.Product;
+using API_Shopping.Models;
 
 namespace API_Shopping.Interfaces
 {
     public interface IProductService
     {
         public Task<object> GetProducts(int page, int pageSize);
-        public Task<object> GetShowProducts(int page, int pageSize, string category);
+        public Task<object> GetCatalogProducts(int page, int pageSize, string category);
         public Task<Product> GetProductById(long id);
         public Task<Product> AddProduct(ProductCreateDTO product);
         public Task<bool> UpdateProduct(long id, ProductUpdateDTO product);
