+++ API_Shopping/API_Shopping/Models/Product.cs
@@ -2,7 +2,7 @@
 {
     public class Product
     {
-        public long ?Id { get; set; }
+        public long Id { get; set; }
         public string Name { get; set; }
         public string Description { get; set; }
         public string Category { get; set; }
